+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Removed
+### Deprecated
 
-- **Breaking:** `APIMixin.patch()` parameter `validate`.
+- `APIMixin.patch()` parameter `validate`. Has no effect and will be removed in a future release.
 
 ## [0.2.0](https://github.com/unioslo/mreg-api/releases/tag/0.2.0) - 2026-04-16